Move known dependencies into language module.
