+++ xee-testrunner/src/language.rs
@@ -1,4 +1,5 @@
 use crate::{
+    dependency::{DependencySpec, KnownDependencies},
     environment::{Environment, XPathEnvironmentSpec, XsltEnvironmentSpec},
     ns::{XPATH_TEST_NS, XSLT_TEST_NS},
     paths::Mode,
@@ -43,4 +44,56 @@ impl Language for XsltLanguage {
     fn mode() -> Mode {
         Mode::Xslt
     }
+}
+
+pub(crate) fn xpath_known_dependencies() -> KnownDependencies {
+    let specs = vec![
+        DependencySpec {
+            type_: "spec".to_string(),
+            value: "XP20+".to_string(),
+        },
+        DependencySpec {
+            type_: "spec".to_string(),
+            value: "XP30+".to_string(),
+        },
+        DependencySpec {
+            type_: "spec".to_string(),
+            value: "XP31+".to_string(),
+        },
+        DependencySpec {
+            type_: "feature".to_string(),
+            value: "higherOrderFunctions".to_string(),
+        },
+        DependencySpec {
+            type_: "xml-version".to_string(),
+            value: "1.0".to_string(),
+        },
+        DependencySpec {
+            type_: "xsd-version".to_string(),
+            value: "1.1".to_string(),
+        },
+    ];
+    KnownDependencies::new(&specs)
+}
+
+pub(crate) fn xslt_known_dependencies() -> KnownDependencies {
+    let specs = vec![
+        DependencySpec {
+            type_: "spec".to_string(),
+            value: "XSLT10+".to_string(),
+        },
+        DependencySpec {
+            type_: "spec".to_string(),
+            value: "XSLT20+".to_string(),
+        },
+        DependencySpec {
+            type_: "spec".to_string(),
+            value: "XSLT30+".to_string(),
+        },
+        DependencySpec {
+            type_: "feature".to_string(),
+            value: "higherOrderFunctions".to_string(),
+        },
+    ];
+    KnownDependencies::new(&specs)
 }
